requirements.txt
faiss-cpu==1.8.0
langchain==0.2.2
langchain-community==0.2.3
openai==1.31.1
streamlit==1.35.0
tiktoken==0.7.0
unstructured-client==0.23.0
PyMuPDF
requests
unstructured
python-magic==0.4.15
psutil
pymongo
bcrypt
python-dotenv